requirements.txt
environs==11.2.1
Django==5.1.3
telebot==0.0.5
pytz==2024.2
python-dotenv==1.1.1
telebot==0.0.5
redis==6.2.0